% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl_crosstab.R
\name{export_formatted_xlsx}
\alias{export_formatted_xlsx}
\title{Export formatted Crosstabs to Excel}
\usage{
export_formatted_xlsx(data, filename, sheet_name = "Crosstab")
}
\arguments{
\item{data}{The data frame from \code{xl_crosstab}.}

\item{filename}{String. Path to the output file (.xlsx).}

\item{sheet_name}{String. Name of the Excel sheet. Default "Crosstab".}
}
\description{
Writes the data frame to an Excel file with text wrapping and top alignment enabled.
}
